requirements.txt
openai==0.27.8
pandas==1.5.3
PyMuPDF==1.22.0
pdfminer==20191125
tiktoken==0.4.0
scikit-learn==1.2.1
scipy==1.10.1
fastapi==0.95.0
uvicorn==0.21.1
faiss-cpu==1.7.3
langchain==0.0.201
urllib3==1.26.16
matplotlib==3.7.1
plotly==5.14.1
python-dotenv==1.0.0
Jinja2==3.1.2
pdf2image==1.16.3
azure-storage-blob==12.16.0
unstructured==0.7.3
pdfplumber==0.9.0
boto3==1.26.162
botocore==1.29.162
pgvector==0.1.8
requests==2.31.0
psycopg2-binary==2.9.6
./library/botocore-1.29.162-py3-none-any.whl
./library/boto3-1.26.162-py3-none-any.whl
